fix(build): stop server-only db chain leaking into client bundle
The supplier evaluation client components imported computeWeightedScore,
scoreToRating and ratingToOutcome from supplier-portal.service, which
transitively imports db.ts -> db-connection-pool -> leak-diagnostics
(uses fs/v8). This dragged Node built-ins into the client bundle and
broke 'next build' with 'Can't resolve fs/v8'.

Extract the pure scoring helpers into a dependency-free module
supplier-evaluation-scoring.ts and import them there from the client
components. The service re-exports them for backward-compatible imports.
